Fix duplicate AssertMessage
